Move all the types to the types folder
