+++ src/hooks/useFiles.ts
@@ -2,7 +2,8 @@ import { useContext, useEffect, useState } from "react";
 import { isEqual } from "lodash";
 
 import FileContext from "../context/FileContext";
-import { FileMetadata } from "../utils/customTypes";
+
+import {FileMetadata} from "../types/fileMetadata";
 
 const useFiles = (): FileMetadata[] => {
   const [validFiles, set] = useState<FileMetadata[]>([]);
