+++ src/types/fileMetadata.tsx
@@ -1,0 +1,11 @@
+export interface FileMetadata {
+    name: string;
+    hash: string;
+    blob: Blob;
+    valid: boolean;
+    keywords?: string[];
+    articles?: number;
+    citations?: number;
+    error?: string;
+    capped?: boolean;
+}
